+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
  - Added Arbutus Cloud and UVic logos to footer and added institutional sign in explanation [#1063](https://github.com/portagenetwork/roadmap/pull/1063)
 
+ - Add handling For Additional User Scenarios in OmniauthCallbacksController [#1074](https://github.com/portagenetwork/roadmap/pull/1074)
+   - 1) Render flash message for users signing in with a "generic name"
+   - 2) Handle attempts to re-link already linked external credentials
+
 ### Fixed
 
  - Fix handling of attrs[:managed] on Org updates [#1062](https://github.com/portagenetwork/roadmap/pull/1062)
@@ -17,6 +21,8 @@
  - Add `upgrade_customization?` Check When Determining Default Template (Funder vs Org Customization) For Dropdown [#1014](https://github.com/portagenetwork/roadmap/pull/1014)
 
  - Clean Up `/contact-us` Page, Fix "Contact us" Footer Link, and Update '/terms' Path Routing [#978](https://github.com/portagenetwork/roadmap/pull/978)
+
+ - Improve OmniauthCallbacksController Tests [#1074](https://github.com/portagenetwork/roadmap/pull/1074)
 
 ### Changed